% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamPattern.TSD.R
\name{beamPattern.TSD}
\alias{beamPattern.TSD}
\title{Returns a function for the beam pattern of an acoustic source. The function beamPattern() has 3 methods depending on the type of the input object 'data'. Inputs are prioritized as (1) functions, (2) NULL, (3) empirical tables, (4) character strings naming functions.}
\usage{
beamPattern.TSD(data, method = c("closest", "linear"))
}
\arguments{
\item{data}{is a list elements defining the beam pattern (parametric or empirical). Names for the elements of the list adopted from read.TSD(). Asterix "*" is either "f", "1" or "2", representing the school (fish), the echo sounder at emission, or the echo sounder at reception:}

\item{method}{is "closest" if the beam pattern value of the closest grid point is to be selected, and "linear" if linear interpolation should be used to extract the beam pattern value (time demanding and only available for 2D grids).}
}
\description{
Returns a function for the beam pattern of an acoustic source. The function beamPattern() has 3 methods depending on the type of the input object 'data'. Inputs are prioritized as (1) functions, (2) NULL, (3) empirical tables, (4) character strings naming functions.
}
\examples{
\dontrun{}

}
