% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularPiston.TSD.R
\name{circularPiston.TSD}
\alias{circularPiston.TSD}
\title{Returns the theoretical far field approximation of the acoustic intensity from a circular baffled source (circular piston) of radius 'b', as a function of incidence data$direle 'data$dire'.}
\usage{
circularPiston.TSD(data, appr = NULL)
}
\arguments{
\item{data}{is a list containing the following elements:}

\item{appr}{is an approximating function used to replace the expression (2*besselJ(ang,1)/ang)^2.}
}
\description{
Returns the theoretical far field approximation of the acoustic intensity from a circular baffled source (circular piston) of radius 'b', as a function of incidence data$direle 'data$dire'.
}
\examples{
\dontrun{}

}
