% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularPiston_ellipticRadius_sidelobefit.TSD.R
\name{circularPiston_ellipticRadius_sidelobefit.TSD}
\alias{circularPiston_ellipticRadius_sidelobefit.TSD}
\title{Returns the theoretical far field approximation of the acoustic intensity from a circular baffled source (circular piston) of radius as an elliptical function of azimuth angle 'ang1' given the radii along the x-axis, kb[,1], and along the y-axis, kb[,2], as a function of incidence angle 'ang2'.}
\usage{
circularPiston_ellipticRadius_sidelobefit.TSD(data, appr = NULL)
}
\arguments{
\item{data}{is a list containing the elements:}

\item{appr}{is an approximating function used to replace the expression (2*besselJ(ang,1)/ang)^2.}
}
\description{
Returns the theoretical far field approximation of the acoustic intensity from a circular baffled source (circular piston) of radius as an elliptical function of azimuth angle 'ang1' given the radii along the x-axis, kb[,1], and along the y-axis, kb[,2], as a function of incidence angle 'ang2'.
}
\examples{
\dontrun{}

}
