% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coltrace.R
\name{coltrace}
\alias{coltrace}
\title{Draws a color spectrum on which the user can assign points with the mouse pointer defining a color trace of desired length.}
\usage{
coltrace(n = 20, ref = NULL, coldens = 20, ...)
}
\arguments{
\item{n}{is the length of the output color vector.}

\item{ref}{is input from a previous session of coltrace(), if the user wishes to adjust an old session.}

\item{coldens}{is the number of color lines to be plotted in the color panel. Large value of 'coldens' is more time demanding but looks nicer.}

\item{...}{methods passed on to sub-functions (not used, but allowing for unused arguments).}
}
\value{

}
\description{
Draws a color spectrum on which the user can assign points with the mouse pointer defining a color trace of desired length.
}
\examples{
\dontrun{}

}
