% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2overlap.R
\name{cor2overlap}
\alias{cor2overlap}
\title{Transforms from correlation values to overlap values used in rexp_MultSines().}
\usage{
cor2overlap(cor)
}
\arguments{
\item{cor}{is a vector or matrix of correlation values.}
}
\value{

}
\description{
Transforms from correlation values to overlap values used in rexp_MultSines().
}
\examples{
\dontrun{}

}
