% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEvent.R
\name{createEvent}
\alias{createEvent}
\title{Creates event skeleton.}
\usage{
createEvent(event, dir, esnm, ow = NULL)
}
\arguments{
\item{event}{The name of the event.}

\item{dir}{The directory in which to put the event.}

\item{esnm}{The names of the acoustic instruments, such as EK60, SU90, ME70 and MS70.}
}
\value{
Paths and name of the events.
}
\description{
Creates event skeleton.
}
