% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBarakatI.R
\name{dBarakatI}
\alias{dBarakatI}
\title{Calculates the probability density function of the total intensity (square of pressure amplitude) from a finite number of sine waves of arbitrary individual amplitudes, given in Equation 64 of Barakat 1974, using the terminology of that paper.}
\usage{
dBarakatI(
  h,
  beta_n2,
  magn = 1,
  N = 100,
  betadistr = c("seq", "flat"),
  max.memory = 1e+09
)
}
\arguments{
\item{h}{is the argument (superimposed intensity).}

\item{beta_n2}{is the vector of individual intensities, or a single number giving the number of significant scatterers.}

\item{N}{is the number of positive roots of the Bessel function of the first kind.}

\item{max.memory}{is the maximum memory occupied by the function before splitting into a loop for each value of 'r'.}
}
\value{

}
\description{
Calculates the probability density function of the total intensity (square of pressure amplitude) from a finite number of sine waves of arbitrary individual amplitudes, given in Equation 64 of Barakat 1974, using the terminology of that paper.
}
\examples{
\dontrun{}

}
