% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.add.noise.R
\name{echoIBM.add.noise}
\alias{echoIBM.add.noise}
\title{Adds noise to an array of sv-values for the ping 'indt', given the noise data present in 'data'.}
\usage{
echoIBM.add.noise(sv, indt, noisetypes = c("bg", "pn", "ms"), data = NULL,
  parlist = list())
}
\arguments{
\item{sv}{is an array dimension [lenb,numb] (or a vector of the same length) of sv-values.}

\item{indt}{is a the time point index, required to pick out the correct seed.}

\item{noisetypes}{is a vector of character strings of length 2, specifying which types of noise to apply to the data:}

\item{data}{is the list of beams inputs as returned from read.TSD. For the treatment of noise the following variables are required:}

\item{parlist}{is a list of input parameters to the noise generation method specified by 'noisetypes'. Use echoIBM_rexp_defaults() to generate defaul values. Must include a seed vector of at least the same length as the maximum indt value to be used.}
}
\value{

}
\description{
Adds noise to an array of sv-values for the ping 'indt', given the noise data present in 'data'.
}
\examples{
\dontrun{}

}
