% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.addRandomness.R,
%   R/echoIBM.fishReaction.R
\name{echoIBM.addRandomness}
\alias{echoIBM.addRandomness}
\alias{echoIBM.fishReaction}
\title{This function adds randomness in the fish positions and to the fish velocities (given by the polarization).}
\usage{
echoIBM.addRandomness(data, grsd_plHS = NULL)

echoIBM.fishReaction(data)
}
\arguments{
\item{data}{A list containing the dynamic fish data, as generated by e.g. \code{\link{echoIBM.generate_dynschool}}.}

\item{grsd_plHS}{A list or data frame linking standard deviations and polarization values. If not given, this is read by grsd_plHS <- read.TSD(system.file("extdata", "grsd_plHS.TSD", package="echoIBM")).}
}
\description{
This function adds randomness in the fish positions and to the fish velocities (given by the polarization).

This function adds randomness in the fish positions and to the fish velocities (given by the polarization).
}
\examples{
\dontrun{}

\dontrun{}

}
