% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.addwarnings.R
\name{echoIBM.addwarnings}
\alias{echoIBM.addwarnings}
\title{Writes warning messages to the end of a echoIBM-dumpfile.}
\usage{
echoIBM.addwarnings(event, warn = NULL, pre = "\\n\\n", post = "")
}
\arguments{
\item{event}{is either the path to a directory containing the dumpfile or the path to the dumpfile itself.}

\item{warn}{is a vector of string with warning messages as returned from catchWarnings() or from warnings2character() or generated in some other way.}

\item{pre}{is a string to be written on the line preceding the warnings.}

\item{post}{is a string to be written on the line following the warnings.}
}
\description{
Writes warning messages to the end of a echoIBM-dumpfile.
}
\examples{
\dontrun{}

}
