% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.bgns.R
\name{echoIBM.bgns}
\alias{echoIBM.bgns}
\title{Expands the background noise into a matrix corresponding to the dimension of the data.}
\usage{
echoIBM.bgns(data, TVG = FALSE, TVG.exp = 2)
}
\arguments{
\item{data}{is a list containing beam configuration, and periodic noise parameters. Speficically the following variables must be included: Beams variables: 'sint', 'lenb'; periodic noise variables: 'pns1', 'pns2', 'pns3', 'acfq', 'harm'; and either 'bgns' or 'numb'}

\item{TVG}{is TRUE if TVG should be added to the periodic noise.}

\item{TVG.exp}{is the exponent of the eamotric spreading of the sound wave, theoretically 2 for Sv and 4 for TS.}
}
\description{
Expands the background noise into a matrix corresponding to the dimension of the data.
}
\examples{
\dontrun{}

}
