% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.default.oneschool.R
\name{echoIBM.default.oneschool}
\alias{echoIBM.default.oneschool}
\title{Performs default handeling of the echoIBM acoustic simulator.}
\usage{
echoIBM.default.oneschool(data, esnm = "MS70", compensated = c("pitch",
  "roll"), max.radius = 0.2, pres = 1e-06, max.cells = 1e+07,
  method = c("closest", "linear"), dumpfile = "dump.txt",
  bptfile = NULL)
}
\arguments{
\item{data}{is a list of data as read by echoIBM.oneschool().}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{compensated}{specifies which rotations are compensated for by the echo sounder. Legal values are "pitch" and "roll", which are both compensated for in the MS70 sonar.}

\item{max.radius}{is the estimated maxumum radius of the transducer elements, used to limit the range over which the radius is calculated if only beam width is given.}

\item{pres}{is the presicion used in the function integrateonsphere() finding the surface integral of the beam pattern of the fish at r=1.}

\item{max.cells}{is the maximum number of cells in the grid used to calculate 'chi'.}

\item{method}{is used in beamPattern.TSD().}

\item{dumpfile}{is the name of the file to which information and warnings about the simulation is written.}

\item{bptfile}{is the name of the file to which 'sllf', 'rad1', 'rad2', 'pbp1' and 'pbp2' is written (use NULL for no writing).}

\item{Static}{variables  NA}

\item{Beam}{pattern  NA}

\item{Configuring}{variables  NA}

\item{Beam}{pattern at emission  NA}

\item{Beam}{pattern at reception  NA}
}
\value{

}
\description{
Performs default handeling of the echoIBM acoustic simulator.
}
\examples{
\dontrun{}

}
