% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.distToPeak.R
\name{echoIBM.distToPeak}
\alias{echoIBM.distToPeak}
\title{Calculates the distance to the nearest peak of the periodic noise for time step 'indt'.}
\usage{
echoIBM.distToPeak(data, indt = NULL)
}
\arguments{
\item{data}{is a list containing beam configuration, and periodic noise parameters. Speficically the following variables must be included: Beams variables: 'sint', 'lenb'; periodic noise variables: 'pns1', 'pns2', 'pns3', 'acfq', 'harm'; and either 'bgns' or 'numb'}

\item{indt}{is the time step index.}
}
\description{
Calculates the distance to the nearest peak of the periodic noise for time step 'indt'.
}
\examples{
\dontrun{}

}
