% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.dump_summary.R
\name{echoIBM.dump_summary}
\alias{echoIBM.dump_summary}
\title{Writes a summary of a variable used by echoIBM() to the dumpfile.}
\usage{
echoIBM.dump_summary(data, dumpfile, type = c("beams", "ctd", "vessel",
  "staticschool", "dynschool", "freqschool", "noise"), append = TRUE)
}
\arguments{
\item{data}{is a variable of which to write summary to the dumpfile.}

\item{dumpfile}{is the path to the file to which to write the information.}

\item{type}{is the type of variable to be dumped.}

\item{append}{is used in write.}
}
\value{

}
\description{
Writes a summary of a variable used by echoIBM() to the dumpfile.
}
\examples{
\dontrun{}

}
