% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.fishInside.R
\name{echoIBM.fishInside}
\alias{echoIBM.fishInside}
\title{Discards fish outside of the sampling region of the sonar/echosounder.}
\usage{
echoIBM.fishInside(dynschool, data, nw = c(r = Inf, az = Inf, el = Inf),
  dumpfile)
}
\arguments{
\item{dynschool}{is a list of the dynamic fish information to be subsetted.}

\item{data}{is a list holding the variables 'esnm', and vessel spesifications.}
}
\value{

}
\description{
Discards fish outside of the sampling region of the sonar/echosounder.
}
\examples{
\dontrun{}

}
