% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.generate_dynschool.R
\name{echoIBM.generate_dynschool}
\alias{echoIBM.generate_dynschool}
\title{Generates the dynamic variables of schools positioned close enough to enter the sampling volume of the sonar.}
\usage{
echoIBM.generate_dynschool(data, t = 1, vesselutim = NULL,
  adds = list(), margin = 0, dumpfile = NULL)
}
\arguments{
\item{data}{is a list containing the necessary data.}

\item{t}{is the current time step.}

\item{vesselutim}{is a vector of the vessel utim points (all time steps in the event)}

\item{margin}{is an extra margin to add, so that the school is detected outside of (positive value) or or inside of (negative value) of the sampling volume.}
}
\value{

}
\description{
Generates the dynamic variables of schools positioned close enough to enter the sampling volume of the sonar.
}
\examples{
\dontrun{}

}
