% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.getFileTypes.R
\name{echoIBM.getFileTypes}
\alias{echoIBM.getFileTypes}
\title{Reads the header of all files in an event and classifies into the following file types: beams (beam configuration), ctd, vessel, pings, school, tsd, other.}
\usage{
echoIBM.getFileTypes(files, ext = c("beams", "ctd", "pings", "school", "tsd",
  "vessel"), key = list(), recursive = TRUE)
}
\arguments{
\item{files}{list of files for which file type should be detemined.}

\item{ext}{Vector of file extensions to test the files against.}

\item{key}{Key varialbe names corresponding to the file types given in \code{ext}.}

\item{recursive}{Used in list.files() when files is a single directory.}
}
\value{

}
\description{
Reads the header of all files in an event and classifies into the following file types: beams (beam configuration), ctd, vessel, pings, school, tsd, other.
}
\examples{
\dontrun{}

}
