% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.interpolate.event.R
\name{echoIBM.interpolate.event}
\alias{echoIBM.interpolate.event}
\title{Runs through segmentation data and interpolates between pings, calculating the probability that voxels in neighboring pings, corresponding to each voxel included in the segmentation mask at a time step, are occupied by the school as well.}
\usage{
echoIBM.interpolate.event(
  event = 1,
  t = 1,
  cruise = 2009116,
  adds = NULL,
  esnm = "MS70",
  type = c("le", "lg", "e", "g", "emax"),
  bw = 0.9,
  thr = NULL,
  kernelthr = 0.1,
  ind = list(-(1:30), NULL),
  range = list(),
  subset = NULL,
  segpar = list(),
  spar = NULL,
  TOV = 0
)
}
\arguments{
\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{t}{is either the number of the ping to be treated, as listed from 1 to the number of pings in the event, or the time point given as a string "yyyymmddHHMMSS.FFF" or "HHMMSS.FFF". Only one time step alowed!}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event.}

\item{adds}{is an optional list of variables overriding the variables in 'data'.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. Currently implemented are "EK60", "ME70", "MS70" and "SH80"/"SX80"/"SH90"/"SX90" (may be given in lover case)}

\item{type}{is the type of interpolation:}

\item{bw}{is the smooting parameter of the interpolation, given as the standard deviation for Gaussian kernel and the base of the exponential decays (bw^x).}

\item{thr}{is the threshold for the interpolated probabilities og school, defining the interpolated segmentation masks.}

\item{kernelthr}{is the threshold below which the kernel is set to 0.}

\item{ind}{is a list of indexes, as typed into the [] of an array, where 0 and NULL denotes all indexes.}

\item{range}{is a list of elements with names matching names of 'data', specifying the range of the corresponding elements. If range is given as a list of 3 elements, xyz}

\item{subset}{is a numeric or logical vector/expression indicating elements or rows to keep. Missing values are taken as false, and subset=0 or subset=NULL indicates no subsetting.}

\item{segpar}{is a list of elements named "bwGp", "lsth"/"rlst", "usth"/"rust", or "sgth"/"sgt0" specifying the parameters of the segmentation data to read.}

\item{spar}{is used in smooth.spline() when the interpolated probabilities of school are smoothed.}

\item{TOV}{is the time offset of the vessel information, discovered by Holmin and Korneliussen in 2013. The default is found in "/Applications/echoIBM/Documentation/Error in yaw MS70/Error in yaw MS70.R".}
}
\description{
Runs through segmentation data and interpolates between pings, calculating the probability that voxels in neighboring pings, corresponding to each voxel included in the segmentation mask at a time step, are occupied by the school as well.
}
\examples{
\dontrun{}

}
