% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.is.schoolfiles.R
\name{echoIBM.is.schoolfiles}
\alias{echoIBM.is.schoolfiles}
\title{Simple function for recognizing and returning school files from a vector of paths to files.}
\usage{
echoIBM.is.schoolfiles(files, names)
}
\arguments{
\item{files}{is a vector of paths to files from which school files are to be extracted.}

\item{names}{is a vector of four character strings representing static and dynamic school valiable names.}
}
\value{

}
\description{
Simple function for recognizing and returning school files from a vector of paths to files.
}
\examples{
\dontrun{}

}
