% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.add.noise.TSD.R, R/echoIBM.merge.R
\name{echoIBM.add.noise.TSD}
\alias{echoIBM.add.noise.TSD}
\alias{echoIBM.merge}
\title{Merges time steps of acoustic data located in the neighbor directory "temp" of the "tsd"-directory given by 'event'.}
\usage{
echoIBM.add.noise.TSD(event, fileind = NULL, beams = NULL, ctd = NULL,
  TVG.in = TRUE, TVG.out = TRUE, TVG.exp = 2, noise = c("nr", "bg",
  "cex"), scls = 1, ow = TRUE, parlist = list(), cores = 1)

echoIBM.merge(event, t = 1, beams = NULL, ctd = NULL, TVG.in = TRUE,
  TVG.out = TRUE, TVG.exp = 2, filesize = 3e+08, noise = c("nr", "bg",
  "cex"), scls = 1, ow = TRUE, parlist = list(), cores = 1,
  reserve = FALSE)
}
\arguments{
\item{event}{is the path to the directory of the simulated files. If the length of 'event' is 2, the second string is the path to the temp-directory holding the files to merge, while the first is the directory in which to put the merged files.}

\item{beams}{is the list of beams inputs as returned from read.TSD. For the treatment of noise the following variables are required:}

\item{ctd}{is the list of ctd inputs as returned from read.TSD.}

\item{TVG.in}{is TRUE if Time Varied Gain compensation is already applied to the input.}

\item{TVG.out}{is TRUE if Time Varied Gain compensation should be applied to the output.}

\item{TVG.exp}{is the exponent in the TVG amplification (usually 2 for Sv and 4 for TS).}

\item{noise}{See echoIBM.add.noise().}

\item{ow}{is TRUE if the user wish to overwrite existing file(s).}

\item{parlist}{is a list of input parameters to the function echoIBM.add.noise(), which generates noise and randomness. See echoIBM.add.noise() for explaination of the possible variables.}

\item{t}{is a vector of the numbers of the pings to be merged, as given in the simulation event. If t=="all", all time steps are merged.}

\item{filesize}{is the maximum size of the merged files.}

\item{reserve}{is FALSE if time steps should not be reserved, requiring that the lengths and dimensions of the variables in the files to merge match exactly. If variable length of the data for each time step, set reserve=TRUE!}

\item{event}{is the path to the directory of the simulated files. If the length of 'event' is 2, the second string is the path to the temp-directory holding the files to merge, while the first is the directory in which to put the merged files.}

\item{beams}{is the list of beams inputs as returned from read.TSD. For the treatment of noise the following variables are required:}

\item{ctd}{is the list of ctd inputs as returned from read.TSD.}

\item{TVG.in}{is TRUE if Time Varied Gain compensation is already applied to the input.}

\item{TVG.out}{is TRUE if Time Varied Gain compensation should be applied to the output.}

\item{TVG.exp}{is the exponent in the TVG amplification (usually 2 for Sv and 4 for TS).}

\item{calibrate.in}{is TRUE if input data is calibrated.}

\item{calibrate.out}{is TRUE if output data should be calibrated.}

\item{noise}{See echoIBM.add.noise().}

\item{ow}{is TRUE if the user wish to overwrite existing file(s).}

\item{parlist}{is a list of input parameters to the function echoIBM.add.noise(), which generates noise and randomness. See echoIBM.add.noise() for explaination of the possible variables.}
}
\value{



}
\description{
Merges time steps of acoustic data located in the neighbor directory "temp" of the "tsd"-directory given by 'event'.

Merges time steps of acoustic data located in the neighbor directory "temp" of the "tsd"-directory given by 'event'.
}
\examples{
\dontrun{}

\dontrun{}

}
