% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.moveSchools.R
\name{echoIBM.moveSchools}
\alias{echoIBM.moveSchools}
\title{Calculates new positions of school specified compactly by the dynamic variables of the school as a whole, given time elapsed.}
\usage{
echoIBM.moveSchools(data, utim)
}
\value{

}
\description{
Calculates new positions of school specified compactly by the dynamic variables of the school as a whole, given time elapsed.
}
\examples{
\dontrun{}

}
