% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.olpt.R
\name{echoIBM.olpt}
\alias{echoIBM.olpt}
\title{Calculates the one step (to the adjacet voxel of the same sampling interval) overlap matrix given by the background noise and the periodic noise for the given time step index. The output is a matrix with the overlap values alont the columns, and one column for each voxel of each beam.}
\usage{
echoIBM.olpt(data, indt = NULL)
}
\arguments{
\item{indt}{is the time index used when constructing the periodic noise.}

\item{adds}{is a list of variables required in the function but not read from the data stored in noise.path(). This is used primarily when creating the file of overlap values, in which case 'crb1' and 'crp1' are needed.}
}
\value{

}
\description{
Calculates the one step (to the adjacet voxel of the same sampling interval) overlap matrix given by the background noise and the periodic noise for the given time step index. The output is a matrix with the overlap values alont the columns, and one column for each voxel of each beam.
}
\examples{
\dontrun{}

}
