% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.oneping.j1.R
\name{echoIBM.oneping.j1}
\alias{echoIBM.oneping.j1}
\title{Simulates one radial sphere of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. One for loop used (for each frequency).}
\usage{
echoIBM.oneping.j1(j, data)
}
\arguments{
\item{j}{is the sampling interval index.}

\item{data}{is the list of data for the simulation, containing all of the following elements:}
}
\value{

}
\description{
Simulates one radial sphere of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. One for loop used (for each frequency).
}
\examples{
\dontrun{}

}
