% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.oneping.j2_old.R,
%   R/echoIBM.oneping.j2_temp.R, R/echoIBM_setB_T.R
\name{echoIBM.oneping.j2}
\alias{echoIBM.oneping.j2}
\alias{echoIBM_setB_T}
\title{Simulates one radial sphere (sample interval) 'j' of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. Two for loops used (for each frequency and each beam).}
\usage{
echoIBM.oneping.j2(j, data)

echoIBM.oneping.j2(j, data)

echoIBM_setB_T(data, direction = 1, beamnr = 1)
}
\arguments{
\item{j}{is the sampling interval index.}

\item{data}{is the list of data for the simulation, containing all of the following elements:}

\item{direction}{Integer: 1 is emision and 2 is reception.}

\item{j}{is the sampling interval index.}

\item{data}{is the list of data for the simulation, containing all of the following elements:
\itemize{
  \item numb
  \item lthesel
  \item transducerposL
  \item thesel
  \item rres
  \item validr
  \item psxf
  \item psyf
  \item pszf
  \item psxv
  \item psyv
  \item pszv
  \item rtzv
  \item rtxv
  \item rtyv
  \item psze
  \item ssil
  \item sigma0mode
  \item uniquek
  \item wavenumber
  \item dira
  \item dire
  \item rad1
  \item bpt1
  \item indi
  \item equalbp_em_re
  \item rad2
  \item bpt2
  \item bptf
  \item lenl
  \item ssif
  \item grsf
  \item absr
  \item fish
  \item asps
  \item epss
  \item epsl
}}

\item{data}{is the list of data for the simulation, containing all of the following elements:
\itemize{
  \item fishdirT
  \item fishposV
  \item dira
  \item dire
  \item wavenumber
  \item lthesel
  \item sllf
  \item indi
  \item bpt*
  \item rad*
  \item pbp*
}}
}
\value{





}
\description{
Simulates one radial sphere (sample interval) 'j' of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. Two for loops used (for each frequency and each beam).

Simulates one radial sphere (sample interval) 'j' of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. Two for loops used (for each frequency and each beam).

Gets the beam pattern values from the transducer, either at emision or reception, depending on the value of direction.
}
\examples{
\dontrun{}

\dontrun{}

\dontrun{}

}
