% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.oneping.oneschool.onesample.R
\name{echoIBM.oneping.oneschool.onesample}
\alias{echoIBM.oneping.oneschool.onesample}
\title{Simulates one radial sphere of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. One for loop used (for each frequency).}
\usage{
echoIBM.oneping.oneschool.onesample(j, data, split = FALSE)
}
\arguments{
\item{j}{is the sampling interval index.}

\item{data}{is the list of data for the simulation, containing all of the following elements:
\itemize{
  \item numb
  \item lthesel
  \item transducerposL
  \item thesel
  \item rres
  \item validr
  \item psxf
  \item psyf
  \item pszf
  \item psxv
  \item psyv
  \item pszv
  \item rtzv
  \item rtxv
  \item rtyv
  \item psze
  \item sigma0mode
  \item uniquek
  \item wavenumber
  \item dira
  \item dire
  \item rad1
  \item bpt1
  \item indi
  \item equalbp_em_re
  \item rad2
  \item bpt2
  \item bptf
  \item lenl
  \item ssif
  \item grsf
  \item absr
  \item fish
  \item asps
  \item epss
  \item epsl
}}

\item{split}{Logical: If TRUE, do the calculation of the transducer beam patterns for each beam, and not for all beams of the same frequency. This saves memory, but increases process time.}
}
\description{
Simulates one radial sphere of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. One for loop used (for each frequency).
}
\examples{
\dontrun{}

}
