% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.oneschool.R
\name{echoIBM.oneschool}
\alias{echoIBM.oneschool}
\title{Simulates echo sounder observations based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school.}
\usage{
echoIBM.oneschool(
  files,
  event,
  t = 1,
  tvessel,
  vesselutim,
  pingsSchool,
  areCompact,
  adds = NULL,
  pingsdir,
  pingsname,
  esnm = NULL,
  TVG.exp = 2,
  compensated = c("pitch", "roll"),
  filesize = 3e+08,
  calibrate = TRUE,
  noise = c("nr", "bg", "ex"),
  mode = c("active", "passive"),
  max.memory = 1e+09,
  ow = TRUE,
  origin = 1,
  dumpfile = "dump.txt",
  dumpsize = 1e+07,
  timedumpfile = "timedump.txt",
  rand.sel = 1,
  scls = 1,
  volsc = c(1, 1, 1),
  method = c("closest", "linear"),
  ask = FALSE,
  parlist = list(),
  bptfile = NULL,
  max.radius = 0.2,
  cores = 1,
  saveBrkt = FALSE,
  msg = FALSE,
  discardOutside = c(r = Inf, az = Inf, el = Inf),
  fishReaction = list()
)
}
\arguments{
\item{files}{is a vector of the paths to the files on which the simulation is based.}

\item{t}{is a vector of the numbers of the pings to be simulated, as listed from 1 to the number of time steps of the simulated school. If t=="all", all time steps are simulated.}

\item{pingsSchool}{is a list of four-column matrices:}

\item{adds}{is an optional list of variables overriding the variables located in the files given by 'files'.}

\item{pingsdir}{is a character string holding the path to the directory in which the .pings files should be put.}

\item{pingsname}{is a character string holding the name of the .pings files, after which the time point of the first ping of each file will be added.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{TVG.exp}{is the exponent of the eamotric spreading of the sound wave, theoretically 2 for Sv and 4 for TS.}

\item{compensated}{specifies which rotations are compensated for by the echo sounder:}

\item{filesize}{is the maximum size of the files, used both for temporal and final files.}

\item{calibrate}{is FALSE if calibration data are to be discarded if present.}

\item{noise}{See echoIBM.add.noise().}

\item{max.memory}{is the maximum amount of memory to allow for the function to occupy. The level of for loops is chosen according to this value, i.e. if the method using only two loops, one for the radial distances and one for the unique frequencies, demands more memory than 'max.memory', the method usint three for loops is chosen.}

\item{ow}{is TRUE if the user wish to overwrite existing file.}

\item{origin}{is a vector of two elements representing the origin of the global coordinate system (G), or the numbering index of the ping regarded as the origin of (G) (ignoring heave so that the x-y-plane of (G) is on the surface of the sea).}

\item{dumpfile}{is the name of the file to which information and warnings about the simulation is written.}

\item{rand.sel}{is a numeric specifying a random selection of the school to use in the simulations. If 'rand.sel' is not >0 and <1 no random selection is done.    If given as a vector of length 2, the second element is regarded as the seed for the random selection.}

\item{method}{is "closest" if the beam pattern value of the closest grid point is to be selected, and "linear" if linear interpolation should be used to extract the beam pattern value (time demanding and only available for 2D grids).}

\item{ask}{is TRUE if the used should be asked to for approval if the memory of the least memory demanding calculation method of the individual radial sampling intervals exceed the memory limit 'max.memory'.}

\item{parlist}{is a list of input parameters to the function echoIBM.add.noise(), which generates noise and randomness. See echoIBM.add.noise() for explaination of the possible variables.}

\item{bptfile}{is the name of the file to which 'sllf', 'rad1', 'rad2', 'pbp1' and 'pbp2' is written (use NULL for no writing).}

\item{cores}{is an integer specifying the number of cores to run the simulations over in parallel (should be lower than the number of cores in the computer).}

\item{saveBrkt}{is TRUE to save the variables 'Brkt' and 'nsig', which contain data only when the number of significant targets is low enough (<=2) for the exact Barakat distribution to kick in instead of the approximation of the exponential distribution.}

\item{msg}{Logical: If TRUE, show messages to the console and to dump files.}

\item{discardOutside}{A vector of three elements specifying the size of the margin around the -3 dB main beam, specifide in units of the range/angular resolution, inside which targets are considered. Using e.g. c(2, 3, 3) will discard targest outside of a volume extending three beam widths around the beam width of the system (and to 2 times the radial range beyond the maximum range). This may save process time, but is not recommended without testing, as higher order side lobes encounter a larger volume than lower order side lobes.}

\item{fishReaction}{A list of variables specifying the fish reaction to the vessel. Must contain the following variables: "magR", "mxrR", "mxdR", "facR". See info.TSD(c("magR", "mxrR", "mxdR", "facR")) for definition of these variables.}
}
\value{

}
\description{
Simulates echo sounder observations based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school.
}
\examples{
\dontrun{}

}
