% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.getPath.R, R/echoIBM.setBeams.R,
%   R/echoIBM.setSchool.R, R/echoIBM.setVessel.R, R/echoIBM.setVessel_old.R
\name{echoIBM.getPath}
\alias{echoIBM.getPath}
\alias{echoIBM.setBeams}
\alias{echoIBM.setSchool}
\alias{echoIBM.setVessel}
\alias{echoIBM.getPath}
\alias{echoIBM.setVessel_old}
\title{Returns a list of the following strings: (1) the path to the event, (2) the event name, (3) the event number, (4) the path to the cruise, and (5) the cruise name.}
\usage{
echoIBM.getPath(starttime = "2015-01-01 00:00:00", origin = NULL,
  heading = NULL, distance = NULL, speed = NULL, duration = NULL,
  nodesLocal = NULL, nodesEarth = NULL, utim = NULL, pingduration = 1,
  heave = 0, rtx = 0, rty = 0, rtz = 0, type = "v")

echoIBM.setBeams(eventDir, eventName = basename(eventDir), utim = NULL,
  esnm = "EK60", data = list(), bmmd = NULL)

echoIBM.setSchool(events, eventName, vessel, surveyRegion = NULL,
  target = "herring", schools = "layer", schoolSize = list(x = 100, y =
  100, z = 20, type = "Weibull", shape = 5), data = list(static = list(),
  dynamic = list()), seed = 0, onlyFirstEvent = FALSE)

echoIBM.setVessel(events, eventName, starttime = "2015-01-01 00:00:00",
  utim = NULL, origin = NULL, heading = NULL, distance = NULL,
  duration = NULL, speed = NULL, nodesLocal = NULL, nodesEarth = NULL,
  pingduration = 1, heave = 0, rtxv = 0, rtyv = 0, rtzv = 0, ...)

echoIBM.getPath(starttime = "2015-01-01 00:00:00", origin = NULL,
  heading = NULL, distance = NULL, speed = NULL, duration = NULL,
  nodesLocal = NULL, nodesEarth = NULL, utim = NULL, pingduration = 1,
  heave = 0, rtx = 0, rty = 0, rtz = 0, type = "v")

echoIBM.setVessel_old(events, eventName, starttime = "2015-01-01 00:00:00",
  origin = NULL, heading = NULL, distance = NULL, speed = NULL,
  duration = NULL, nodesLocal = NULL, nodesEarth = NULL, utim = NULL,
  numt = 1, pingduration = 1, heave = 0, rtxv = 0, rtyv = 0,
  rtzv = NULL, ...)
}
\arguments{
\item{starttime}{The start time of the event, i.e., the time of the first ping/vessel position, either given as a time object or a string such as "2015-01-01 00:00:00".}

\item{origin}{The origin of the event, i.e., a vector of two elements giving the longitude and latitude of the first ping.}

\item{heading}{The heading of the vessel in radians counter clockwise from North, either given as a single numeric, or as a vector of headings associated with vessel track segments.}

\item{distance}{The distance associated with each vessel track segment.}

\item{speed}{The speed of the vessel in knots associated with each vessel track segment.}

\item{duration}{The duration of each vessel track segment.}

\item{nodesLocal}{The nodes defining the vessel track segments, given as local nodes relative to the origin (can include the origin c(0, 0))}

\item{nodesEarth}{The nodes defining the vessel track segments, given as global nodes in the coordinate system of the earth relative to the origin (can include the origin c(0, 0)).}

\item{utim}{The UNIX time points of the event. If not given, the starttime and pingduration will define the UNIX time points.}

\item{pingduration}{The duration of the pings in seconds. Currently only one single fixed value is allowed.}

\item{heave}{The heave of the vessel in meters, either given as a single value, a vector of length equal to the number of vessel track segments, or alternatively a funciton of the number of time steps such as rnorm() or the more appropriate function(x) {set.seed(x); runif(x)}, which sets the seed as the number of time steps.}

\item{eventName}{The name of the simulation event, used in the file name of the vessel file.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{rtxv}{The pitch of the vessel in radians positive for uppwards pitch (bow lifting). See info.TSD("rtxv")}

\item{rtyv}{The roll of the vessel in radians positive for starboard side tilted down. See info.TSD("rtyv")}

\item{rtzv}{The orientation of the vessel in radians counter clockwise, IN ADDITION to the heading of the vessel specified in \code{heading}. See info.TSD("rtzv")}

\item{...}{is used in agrep() for locating events based on approximate string matching.}

\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{dir.type}{is the name of the directory holding the data files (usually one of "tsd" and "raw")}

\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{dir.type}{is the name of the directory holding the data files (usually one of "tsd" and "raw")}

\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{dir.type}{is the name of the directory holding the data files (usually one of "tsd" and "raw")}

\item{...}{is used in agrep() for locating events based on approximate string matching.}

\item{...}{Further parameters passed to or from other methods.}

\item{event}{is the identifier of the event, either given as the number of the event, a string contained in the name of the event, or the path of the event directory.}

\item{cruise}{is either the idenfication number of the cruise, given as specified by the IMR (yyyynnn), or the path to the directory containing the event to be read.}

\item{esnm}{is the name of the acoustical instrument, given as a four character string. See sonR_implemented() for the implemented systems. May be given in 'data', and in lower case.}

\item{dir.type}{is the name of the directory holding the data files (usually one of "tsd" and "raw")}

\item{...}{is used in agrep() for locating events based on approximate string matching.}
}
\value{









}
\description{
Returns a list of the following strings: (1) the path to the event, (2) the event name, (3) the event number, (4) the path to the cruise, and (5) the cruise name.

Returns a list of the following strings: (1) the path to the event, (2) the event name, (3) the event number, (4) the path to the cruise, and (5) the cruise name.

Returns a list of the following strings: (1) the path to the event, (2) the event name, (3) the event number, (4) the path to the cruise, and (5) the cruise name.

The vessel information can be given in 4 ways: 
\itemize{
\item nodesEarth + speed
\item origin + nodesLocal + speed
\item origin + heading + distance + speed
\item origin + heading + duration + speed
}

Returns a list of the following strings: (1) the path to the event, (2) the event name, (3) the event number, (4) the path to the cruise, and (5) the cruise name.
}
\examples{
\dontrun{}

\dontrun{}

\dontrun{}

\dontrun{}

\dontrun{}

}
