% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.w2scale.R
\name{echoIBM.w2scale}
\alias{echoIBM.w2scale}
\title{Converts from radial overlap values 'w' in sonar and echosounder systems to factors 'scale' used for calibrating simulations made by echoIBM() using the method of superimposing multiple sine waves for generation of correlated exponentially distributed values.}
\usage{
echoIBM.w2scale(w)
}
\arguments{
\item{w}{is the radial overlap value, specifically the duration of the sine waves used to generate the correlated exponentially distributed values, usually calculated as the ratio of pulselength and sampling interval duration multiplied by the factor 3/4 (experienced from previous simulations to be suitable for the MS70 sonar and the SX90 sonar).}
}
\description{
Converts from radial overlap values 'w' in sonar and echosounder systems to factors 'scale' used for calibrating simulations made by echoIBM() using the method of superimposing multiple sine waves for generation of correlated exponentially distributed values.
}
\examples{
\dontrun{}

}
