% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM_rexp_MultSines.R
\name{echoIBM_rexp_MultSines}
\alias{echoIBM_rexp_MultSines}
\title{Calls the function rexp_MultSines(), which generates correlated vectors of autocorrelated exponential/Barakat variables. Only one time step is generated.}
\usage{
echoIBM_rexp_MultSines(parlist = list(), data = list(), Npre = 1000,
  shape = 1)
}
\arguments{
\item{parlist}{is a list of input parameters:}

\item{data}{is a list of beam configuration data, used to extract defaults if 'parlist' is not given.}
}
\value{

}
\description{
Calls the function rexp_MultSines(), which generates correlated vectors of autocorrelated exponential/Barakat variables. Only one time step is generated.
}
\examples{
\dontrun{}

}
