% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.setup.R
\name{echoIBM.setup}
\alias{echoIBM.setup}
\title{***DECRIPTION MISSING***}
\usage{
echoIBM.setup(eventName, dir, t = "all", pre = TRUE, vessel = list(),
  schoolStatic = list(), schoolDynamic = list(), ctd = list(),
  beams = list(), noise = list(), calibration = list(), addnoise = TRUE,
  passive = FALSE, cores = 1)
}
\value{

}
\description{
***DECRIPTION MISSING***
}
\examples{
\dontrun{}

}
