% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etaCompression.R
\name{etaCompression}
\alias{etaCompression}
\title{The hydrostatic compression factor of a swim bladder. Two cases are supported, combining the compression of length and width, given by L(z) = L(0)(1-(rho*g*z)/p0)^gammaL and W(z) = W(0)(1-(rho*g*z)/p0)^gammaW: (1) Area: The hydrostatic compression factor of the cross sectional area of the swim bladder is found by multiplication of L and W, resulting in (1-(rho*g*z)/p0)^(gammaL+gammaW). (2) The hydrostatic compression factor of the oblongness is found by the ratio L/W, giving (1-(rho*g*z)/p0)^(gammaL-gammaW)}
\usage{
etaCompression(data = list(), type = c("area", "obln"))
}
\arguments{
\item{data}{is a list containing the elements "rho0", specifying the mass density of the sea, "gacc", representing the gravitational accelleration, "pszf" which is the z-position of the fish in (G), "hpr0", giving the hydrostatic pressure at sea surface (equal to air pressure at the sea surface) and "gaml" and "gamw" representing the depth compression parameters of the length and the width of the swim bladder model.}

\item{type}{is a character string representing the type of the compression factor, where "area" denotes compression of the cross sectional area of the swin bladder and "obln" corresponds to changes in the oblongness of the swin bladder (see DESCRIPTION).}
}
\description{
The hydrostatic compression factor of a swim bladder. Two cases are supported, combining the compression of length and width, given by L(z) = L(0)(1-(rho*g*z)/p0)^gammaL and W(z) = W(0)(1-(rho*g*z)/p0)^gammaW: (1) Area: The hydrostatic compression factor of the cross sectional area of the swim bladder is found by multiplication of L and W, resulting in (1-(rho*g*z)/p0)^(gammaL+gammaW). (2) The hydrostatic compression factor of the oblongness is found by the ratio L/W, giving (1-(rho*g*z)/p0)^(gammaL-gammaW)
}
\examples{
\dontrun{}

}
