% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etaOmega.R
\name{etaOmega}
\alias{etaOmega}
\title{The aspect factor (cross sectional area relative to the maximum cross sectional area) of a cylindrical object of rounded ends, a prolate spheroid or a point source (circular). See the documentation of echoIBM for details (Section 3.1.4).}
\usage{
etaOmega(data, pbpf = c("ls", "pr", "ps"))
}
\arguments{
\item{data}{is a list containing the elements "obln", specifying the oblongness of the object (ratio of length and thickness) and "transducerposL", giving the position of the transducet in the coordinate system of the target in spherical coordinates. Also "pbpf" may be given in 'data'.}

\item{pbpf}{is a string naming the theoretical acoustical model of the target (may be included in 'data').}
}
\value{

}
\description{
The aspect factor (cross sectional area relative to the maximum cross sectional area) of a cylindrical object of rounded ends, a prolate spheroid or a point source (circular). See the documentation of echoIBM for details (Section 3.1.4).
}
\examples{
\dontrun{}

}
