% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_eps.R
\name{fun_eps}
\alias{fun_eps}
\title{Returns a function for the echo ability 'eps*' equal to the optimal acoustic cross sectional area of a target divided by the length of the target to the desired power (usually squared). The function fun_eps() has 2 methods depending on the type of the input object 'data'.}
\usage{
fun_eps(data = NULL, Nl = 1, method = c("closest", "linear",
  "constant"))
}
\arguments{
\item{data}{may be of 2 different types.}

\item{Nl}{is the number of targets from which the echo is simulated.}

\item{method}{defines the interpolation method.}
}
\value{

}
\description{
Returns a function for the echo ability 'eps*' equal to the optimal acoustic cross sectional area of a target divided by the length of the target to the desired power (usually squared). The function fun_eps() has 2 methods depending on the type of the input object 'data'.
}
\examples{
\dontrun{}

}
