% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.grid_phi.R
\name{get.grid_phi}
\alias{get.grid_phi}
\title{Calculates the grid of elevation angle phi, used in integrateonsphere() to make all cells have the same area.}
\usage{
get.grid_phi(from = 0, to = pi, area = 100)
}
\arguments{
\item{from}{is the lower limit (of phi) of the integral.}

\item{to}{is the upper limit (of phi) of the integral.}

\item{area}{is the length of the grid vector if length(area)==1, and the relative sizes of the cells, if a vector of length > 1 is given (less useful).}
}
\value{

}
\description{
Calculates the grid of elevation angle phi, used in integrateonsphere() to make all cells have the same area.
}
\examples{
\dontrun{}

}
