% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.transducerRadius2.TSD.R
\name{get.transducerRadius2.TSD}
\alias{get.transducerRadius2.TSD}
\title{Calculates the radii of circular transducer elements based on -3 dB beam angles.}
\usage{
get.transducerRadius2.TSD(data)
}
\arguments{
\item{beamwidth}{is a vector of -3 dB beamwidths for the beams given in radians (the angle between the directions in which the acoustic intensity is half of the maximum acoustic intensity).}

\item{beampattern}{is the beam pattern function having angle in radians as its first argument and the product of wave number and transducer element radius as its second argument, e. g. circularPiston(ang,kb).}

\item{wavenumber}{is a vector of wave numbers for the beams.}

\item{max.radius}{is an expected maximum radius which should be low enough to avoid exhaustive or inaccurate calculations.}

\item{twoway}{is TRUE if the beampattern is to be squared, corresponding to a two way beam forming. I.e., if the beam widths are given as two-way angles bwtl=4 degrees, the one way beam width is actually 4*sqrt(2)=5.656854, and twoway should be TRUE to account for this.}
}
\value{

}
\description{
Calculates the radii of circular transducer elements based on -3 dB beam angles.
}
\examples{
\dontrun{}

}
