% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqBarakat.R
\name{getSeqBarakat}
\alias{getSeqBarakat}
\title{Generates 'beta_n' used in dBarakatA(), pBarakatA(), dBarakatI(), pBarakatI(), rBarakatA(), rBarakatI().}
\usage{
getSeqBarakat(beta_n, betadistr = c("seq", "flat"))
}
\arguments{
\item{beta_n}{is the vector of individual pressure amplitudes, or a single number giving the number of significant scatterers.}

\item{betadistr}{is an identifier specifying how to generate 'beta_n' in the case that 'beta_n' is a single numeric.}
}
\description{
Generates 'beta_n' used in dBarakatA(), pBarakatA(), dBarakatI(), pBarakatI(), rBarakatA(), rBarakatI().
}
\examples{
\dontrun{}

}
