% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_simple.R
\name{hist_simple}
\alias{hist_simple}
\title{Histogram without plot, faster.}
\usage{
hist_simple(x, breaks = 10)
}
\arguments{
\item{x}{is a vector.}

\item{breaks}{is the number of breaks (in which case pretty() is used), or a vector of breaks.}
}
\value{

}
\description{
Histogram without plot, faster.
}
\examples{
\dontrun{}

}
