% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateonsphere.R
\name{integrateonsphere}
\alias{integrateonsphere}
\title{Numerical calculation of the surface integral of the function 'fun' on (parts of) the unit sphere.}
\usage{
integrateonsphere(
  fun,
  ndim = 2,
  theta = c(0, 2 * pi),
  phi = c(0, pi),
  R = 1,
  pres = 1e-06,
  l = c(100, 100),
  incriment = 2,
  equal.size = TRUE,
  max.cells = 1e+06,
  print = TRUE
)
}
\arguments{
\item{fun}{is the input function to integrate. Must take a column matrix of theta- and phi-values, or a vector of phi-values as input, and return a vector for the paris of 'theta' and 'phi'.}

\item{ndim}{is the number of dimensions of the input to 'fun'.}

\item{theta}{is a vector of two elements representing the lower and upper bounds of theta (azimuth angle). Only needed if ndim==2.}

\item{phi}{is a vector of two elements representing the lower and upper bounds of the phi (elevation angle).}

\item{R}{is the radius of the sphere.}

\item{pres}{is the desired presition of the integration.}

\item{l}{is the initial lengths of the theta and phi grid.}

\item{incriment}{is the factor to multiply the lengths of the grids by.}

\item{equal.size}{is TRUE if all cells are to have equal size (speeds up function) and FALSE if the anglular incriment is to be equal.}

\item{max.cells}{is the maximum number of cells in the grid.}

\item{print}{is TRUE if the improvement of the iteration is to be printed.}
}
\description{
Numerical calculation of the surface integral of the function 'fun' on (parts of) the unit sphere.
}
\examples{
\dontrun{}

}
