% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineSource.TSD.R
\name{lineSource.TSD}
\alias{lineSource.TSD}
\title{Returns the theoretical far field approximation of the acoustic intensity from a line source of length 'L', as a function of incidence data$direle 'data$dire'.}
\usage{
lineSource.TSD(data)
}
\arguments{
\item{data}{is a list containing the elements 'dire' (elevarion angle in radians [0,pi/2]) and 'wnsz' (product of wave number and size), and optionally 'dira' (azimuth angle in radians [0,pi/2]), 'indi' (index of the fish / transducer element) and 'obln' (oblongness of the ).}
}
\value{

}
\description{
Returns the theoretical far field approximation of the acoustic intensity from a line source of length 'L', as a function of incidence data$direle 'data$dire'.
}
\examples{
\dontrun{}

}
