% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ebpf.R
\name{merge_ebpf}
\alias{merge_ebpf}
\title{Merges empirical beam patterns located in separate TSD-files. If duplicated 'grsf' values are present the first is allways chosen.}
\usage{
merge_ebpf(files, con = NULL)
}
\arguments{
\item{files}{is a vector of strings representing the paths to the files to merge, or a single string representing the directory in which the files are located.}
}
\value{

}
\description{
Merges empirical beam patterns located in separate TSD-files. If duplicated 'grsf' values are present the first is allways chosen.
}
\examples{
\dontrun{}

}
