% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsignif.R
\name{nsignif}
\alias{nsignif}
\title{Returns an estimate of the number of significant values in 'x', where the significance of the values are measured as the fractional magnitude relative to the maximum value.}
\usage{
nsignif(x, na.rm = TRUE)
}
\arguments{
\item{x}{is a numeric vector.}
}
\description{
Returns an estimate of the number of significant values in 'x', where the significance of the values are measured as the fractional magnitude relative to the maximum value.
}
\examples{
\dontrun{}

}
