% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBarakatA.R
\name{pBarakatA}
\alias{pBarakatA}
\title{Calculates the cummulative probability function of the pressure amplitude from a finite number of sine waves of arbitrary individual amplitudes, given (for the upper probability) in Equation 59 of Barakat 1974, using the terminology of that paper.}
\usage{
pBarakatA(r_0, beta_n, magn = 1, N = 100, betadistr = c("seq", "flat"),
  max.memory = 1e+09)
}
\arguments{
\item{r_0}{is the argument (superimposed pressure amplitude).}

\item{beta_n}{is the vector of individual pressure amplitudes, or a single number giving the number of significant scatterers.}

\item{N}{is the number of positive roots of the Bessel function of the first kind.}

\item{max.memory}{is the maximum memory occupied by the function before splitting into a loop for each value of 'r'.}
}
\value{

}
\description{
Calculates the cummulative probability function of the pressure amplitude from a finite number of sine waves of arbitrary individual amplitudes, given (for the upper probability) in Equation 59 of Barakat 1974, using the terminology of that paper.
}
\examples{
\dontrun{}

}
