% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointSource.TSD.R
\name{pointSource.TSD}
\alias{pointSource.TSD}
\title{Returns omnidirectional directional factor of dimension equal to the first input.}
\usage{
pointSource.TSD(data)
}
\arguments{
\item{data}{is a list containing the elements 'dire' (elevarion angle in radians [0,pi/2]) and 'wnsz' (product of wave number and size), and optionally 'dira' (azimuth angle in radians [0,pi/2]), 'indi' (index of the fish / transducer element) and 'obln' (oblongness of the ).}
}
\value{

}
\description{
Returns omnidirectional directional factor of dimension equal to the first input.
}
\examples{
\dontrun{}

}
