% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol2sd.R
\name{pol2sd}
\alias{pol2sd}
\title{Gets the standard deviation to apply to the function ARsim_school().}
\usage{
pol2sd(pol, speed = 1, dt = 1, gamma = 0.8, type = "h")
}
\arguments{
\item{pol}{is the desired polarization value, either given as the Huth measure (mean angle deviation between individual and group heading) or the Couzin measure.}

\item{speed}{is the speed of the fish.}

\item{dt}{is the difference between time steps.}

\item{type}{see pol.school().}
}
\value{

}
\description{
Gets the standard deviation to apply to the function ARsim_school().
}
\examples{
\dontrun{}

}
