% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repToExtreme.R
\name{repToExtreme}
\alias{repToExtreme}
\title{Repeats all elements of the list \code{x} to the}
\usage{
repToExtreme(x, ndim = 2, skip.len = 1)
}
\arguments{
\item{x}{A list of arrays.}

\item{ndim}{The number of dimensions of the arrays to repeat to the maximum dimension.}

\item{skip.len}{The length of arrays to leave be.}
}
\value{

}
\description{
Repeats all elements of the list \code{x} to the
}
\examples{
\dontrun{}

}
