% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp_MultSines.R
\name{rexp_MultSines}
\alias{rexp_MultSines}
\title{Returns an array of correlated vectors of exponentially distributed variables, where each vector has autocorrelation resulting from superimposed sine waves of equal length, frequency and amplitude, but with uniformly distributed starting point. The values are generated by dividing the time into intervals, and extracting the central peak of the superimposed wave for each sampling interval. Correlation between vectors is acomplished by also including sine waves from other vectors in the superimposed wave, to a degree given by 'olpn'. For low numbers of sine waves the square of the probability distribution described by Barakat 1974 emerges, but strong deviations from the exponential distribution requires very low values of 'L' and 'w' and few elements in 'olpn', as the Rayleigh distribution kicks in as low as around 4 individual waves. WARNING: The expectation of the exponential variables is dependent on the parameters used, and for the default setting it is equal to 7.350175.}
\usage{
rexp_MultSines(J = 1324, I = 25, L = 3, N = 40, P = 10, w = 3,
  olpn = c(0.5, 1, 0.5), shape = 1, mean = 1, seed = NULL)
}
\arguments{
\item{J}{is the length of the beams.}

\item{I}{is the number of beams.}

\item{L}{is the number of fish in each voxel.}

\item{N}{is the number of sample points in each voxel (N=4*P, the sampling points are expected to end up close to peaks along the sine waves).}

\item{P}{is the is the frequency of the sine waves, meaning the number of periods per voxel.}

\item{w}{is the length of the sine waves in units of the time intervals constituting the voxels.}

\item{olpn}{is a vector, matrix or array of correlations, representing constant correlation for all voxels, variable correlation between beams, and variable correlation for all voxels.}

\item{shape}{is is used to transform to a Weibull variable of the desired shape and scale=1.}

\item{seed}{is the seed of the function.}
}
\value{

}
\description{
Returns an array of correlated vectors of exponentially distributed variables, where each vector has autocorrelation resulting from superimposed sine waves of equal length, frequency and amplitude, but with uniformly distributed starting point. The values are generated by dividing the time into intervals, and extracting the central peak of the superimposed wave for each sampling interval. Correlation between vectors is acomplished by also including sine waves from other vectors in the superimposed wave, to a degree given by 'olpn'. For low numbers of sine waves the square of the probability distribution described by Barakat 1974 emerges, but strong deviations from the exponential distribution requires very low values of 'L' and 'w' and few elements in 'olpn', as the Rayleigh distribution kicks in as low as around 4 individual waves. WARNING: The expectation of the exponential variables is dependent on the parameters used, and for the default setting it is equal to 7.350175.
}
\examples{
\dontrun{}

}
