% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounded.rect.R
\name{rounded.rect}
\alias{rounded.rect}
\title{Draws a rounded rectangle.}
\usage{
rounded.rect(origin = 0, w = 1, h = 1, r = 0.1, n = 100,
  plot = TRUE, adjust = FALSE, by = "hw", ...)
}
\arguments{
\item{origin}{is the origin of the rounded rectangle.}

\item{w}{is the width of the rounded rectangle.}

\item{h}{is the height of the rounded rectangle.}

\item{r}{is the radius of the rounded corners as a fraction on 'w'.}

\item{n}{is the number of points along the rounded corners of the rounded rectangle.}

\item{plot}{is whether the rounded rectangle is to be plotted or not.}

\item{adjust}{is TRUE if the rounded corners should appear as circles regardless of the frame dimensions.}

\item{by}{is along which of width, height or both, the adjustment for reshaped window dimensions of the rounded rectangles are to be done.}

\item{...}{are parameters like 'col' and 'border' used in polygon().}
}
\value{

}
\description{
Draws a rounded rectangle.
}
\examples{
\dontrun{}

}
