% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.path.vessel.cm.R
\name{set.path.vessel.cm}
\alias{set.path.vessel.cm}
\title{Wrapper for the method of drawing vessel positions based on school centres of mass, given as an option to the fundtion echoIBM (specifying path=TRUE). set.path.vessel.cm() uses the function set.path.vessel() to select the vessel positions. The input variables in this function are especially selected for the single purpose of being used in echoIBM(), and are thus not user friendly for any othr use.}
\usage{
set.path.vessel.cm(schooldirs, dynschoolfiles, recycle, t, tvessel, numt,
  smooth, margin, scan.volume, beams, ctd, rph, event, eventname, pathnr)
}
\arguments{
\item{schooldirs}{is a vector of the directories of the schools to be simulated.}

\item{dynschoolfiles}{is a list of the dynamic school files for the schools to be simulated.}

\item{recycle}{see echoIBM(). Here, if 'recycle' is different from FALSE or does not contains zeros, the first time step is selected.}

\item{t}{is the vector of the time steps for which to draw the vesse movement.}

\item{tvessel}{is the UNIX time values to be written to the ".vessel" file. If tvessel==NULL the school time points are used.}

\item{numt}{is the number of time steps for which to select vessel positions.}

\item{smooth}{is the function used for smoothing the vessel path.}

\item{margin}{is a vector of the margins on either side of the span of 'x' (recycled if not of length 4).}

\item{scan.volume}{is TRUE if the acoustic observation volume schould be scanned for the positions of the fish.}

\item{beams}{is a list of the beam configuration.}

\item{ctd}{is a list of the CTD-specification of the sea.}

\item{rph}{is a matrix of two columns of length 3 representing the mean (column 1) and the standard deviation (column 2) of the roll values (rtxv), pitch values (rtyv) and heave values (przv) of the vessel, used in gaussian simulation of the roll, pitch and heave values.}

\item{event}{is a vector of strings representing the paths to the directories and files containing the necessary informtation for the simulation.}

\item{eventname}{is the name of the main directory of the simulation, in which to store the simulate acoustical files.}

\item{pathnr}{is the number of the vessel path drawn, to be used in the name of the ".vessel" file. If a new ".vessel" file is to be written, 'pathnr' must be different than the pathnr of existing files.}
}
\value{

}
\description{
Wrapper for the method of drawing vessel positions based on school centres of mass, given as an option to the fundtion echoIBM (specifying path=TRUE). set.path.vessel.cm() uses the function set.path.vessel() to select the vessel positions. The input variables in this function are especially selected for the single purpose of being used in echoIBM(), and are thus not user friendly for any othr use.
}
\examples{
\dontrun{}

}
