% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.path.vessel.R
\name{set.path.vessel}
\alias{set.path.vessel}
\title{Interactive function for selecting the path and speed of an object based on reference points 'x', intended for use in echoIBM().}
\usage{
set.path.vessel(
  x,
  n = nrow(x),
  t = NULL,
  ref = NULL,
  maxspeed = 15,
  margin = 500,
  maxn = 100,
  path.col = "blue",
  grid = TRUE,
  smooth = "spline"
)
}
\arguments{
\item{x}{are the reference points.}

\item{n}{and 't' are the numbers of points and the times for the points along the path respectively. The combination of 'n' and 't', and whether the user selects speed values interactively, defines the mode of the output. There are 7 modes of the output, as liste below (lt=length(t)):}

\item{ref}{is a list of reference information from a previous set.path.vessel session (given as the output from set.path.vessel()). The points of ref will be plotted in grey, to form a visual reference for the user.}

\item{margin}{is a vector of the margins on either side of the span of 'x' (recycled if not of length 4).}

\item{maxn}{maximum number of points entered (to prevent "getting stuck" in the function).}

\item{path.col}{color of the entered path.}

\item{grid}{is TRUE if a 100 meter grid is to be plotted.}

\item{smooth}{has one possible value "spline", smooths the x-values and the y-values of the path separately using the spline function.}
}
\description{
Interactive function for selecting the path and speed of an object based on reference points 'x', intended for use in echoIBM().
}
\examples{
\dontrun{}

}
