% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonR_implemented_ind.R
\name{sonR_implemented_ind}
\alias{sonR_implemented_ind}
\title{Implemented ind strings in echoIBM and realted functions}
\usage{
sonR_implemented_ind(ind)
}
\arguments{
\item{ind}{see ind.expand()}
}
\description{
Implemented ind strings in echoIBM and realted functions
}
