% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.line.R
\name{track.line}
\alias{track.line}
\title{Returns a list of x-values and y-values of points defined by the fractions 'w' along the partial linear path defined by the points (x,y). Cumulative lengths of the path are also returned.}
\usage{
track.line(
  x,
  y = NULL,
  w = NULL,
  dw = NULL,
  speed = NULL,
  start = NULL,
  alongx = FALSE,
  normalized = FALSE,
  list.out = TRUE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{is the input x-variable.}

\item{y}{is the input y-variable.}

\item{w}{are the positions along the path (x,y). If 'w' is a single integer, it is set to seq(0,1,length.out=w). If 'speed' is given 'w' is interpreted as time values.}

\item{dw}{are the piecewise lengths along the path (x,y), used only if w!=NULL. If 'dw' has length>1 it overrides 'w' by w=cumsum(dw).}

\item{speed}{are the speed values along the path (x,y), or along the x-axis if alongx==TRUE. If given it overrides 'w'.}

\item{start}{is a vector of two values representing the start point if 'x' and 'y' represent angles and lengths. If 'start' is given, 'x' and 'y' are interpreted as angles and lengths.}

\item{alongx}{is TRUE if tracking is to be done along the x-axis, and not along the path (x,y).}

\item{normalized}{is TRUE if the positions given by 'w' (and 'speed') are or should be normalized to [0,1], where 1 refers to the end of the path.}

\item{list.out}{is TRUE if the output should be a data frame with the names "x", "y" and "l". Else a matrix with the same names is returned.}

\item{plot}{is TRUE if a simple plot is to be drawn.}

\item{...}{are parameters to be passed on to plot() if plot==TRUE.}
}
\description{
Returns a list of x-values and y-values of points defined by the fractions 'w' along the partial linear path defined by the points (x,y). Cumulative lengths of the path are also returned.
}
\examples{
\dontrun{}

}
