% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol.sph.R
\name{vol.sph}
\alias{vol.sph}
\title{Calculates the (volume) size of the spherical volume segment defined by 'r', 'theta' and 'phi'.}
\usage{
vol.sph(r = c(0, 1), theta = c(0, 2 * pi), phi = c(0, pi),
  var = c("volx", "harx"))
}
\arguments{
\item{r}{is range of radial coordinates, given as a vector of interval values (in which case length(r)-1 spherical segemnts are treated) or as a two column matrix where the first column represents the lower limit of the sperical segments, and the second column represents the upper limit.}

\item{theta}{is the range of azimuth angles in radians, given in the same way as 'r'.}

\item{phi}{is the range of elevation angles in radians, given in the same way as 'r.}

\item{var}{is a vector of the variables to return. Currently implemented are "volx" for volumes of the voxels and "harx" for horizontal area of the voxels.}
}
\value{

}
\description{
Calculates the (volume) size of the spherical volume segment defined by 'r', 'theta' and 'phi'.
}
\examples{
\dontrun{}

}
