% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.ssif.R
\name{write.ssif}
\alias{write.ssif}
\title{Calculates and (optionally) writes to file the spherical surface integral of a beam pattern from acoustic fish models at radius 1 meter, for the grid of the product 'kL' of wavenumber and size of the acoustic model.}
\usage{
write.ssif(x = NULL, outfile = NULL, pres = 1e-06,
  max.cells = 1e+06, onlyssif.out = TRUE)
}
\arguments{
\item{x}{is either a list of the data ('grsf' (grid of the product of wavenumber and size of the acoustic model) and 'ebpf' (empirical beam pattern table) and at least one of 'graf' (grid azimuth angle) and 'gref' (grid elevation angle)) or a path to the TSD file holding the empirical beam pattern.}

\item{outfile}{is the path to the file to which the data should be written, FALSE if no data should be written to file, or if length(outfile) == 0 the default file name which is the input file added "_ssif" to the name is used (only applied in the case that 'x' is a TSD-file).}

\item{pres}{is the desired presition of the integration (used in integrateonsphere()).}

\item{max.cells}{is the maximum number of cells in the grid (used in integrateonsphere()).}

\item{onlyssif.out}{is TRUE if the output should be only 'ssif', not the entire list read, added 'ssif'.}
}
\value{

}
\description{
Calculates and (optionally) writes to file the spherical surface integral of a beam pattern from acoustic fish models at radius 1 meter, for the grid of the product 'kL' of wavenumber and size of the acoustic model.
}
\examples{
\dontrun{}

}
