% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2ab.R
\name{xy2ab}
\alias{xy2ab}
\title{Parametrizes the piecewise linear trace given by the input. For vertical lines, $Inf$ is returned.}
\usage{
xy2ab(x, y = NULL, ang.out = FALSE)
}
\arguments{
\item{x}{and 'y' are the x-coordinates and y-coordinates of the input trace, given as separate vectors 'x' and 'y', or as a list or matrix 'x' holding both the x-coordinates and the y-coordinates, or if y==NULL 'x' is interpreted as 'y' and x=seq_along(x).}

\item{ang.out}{is TRUE if angles are to be returned instead of slope.}
}
\value{

}
\description{
Parametrizes the piecewise linear trace given by the input. For vertical lines, $Inf$ is returned.
}
\examples{
\dontrun{}

}
