% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularPiston_appr.R
\docType{data}
\name{besselJ1_appr_angmax}
\alias{besselJ1_appr_angmax}
\title{Function used for approximating the beam pattern of the beams}
\format{An object of class \code{numeric} of length 1.}
\usage{
besselJ1_appr_angmax
}
\arguments{
\item{x}{is a vector.}
}
\value{

}
\description{
Function used for approximating the beam pattern of the beams
}
\examples{
\dontrun{}

}
\keyword{datasets}
