% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.extraxtRandSel.R,
%   R/echoIBM.fishInside.R, R/echoIBM.fishInside2.R
\name{echoIBM.extraxtRandSel}
\alias{echoIBM.extraxtRandSel}
\alias{echoIBM.fishInside}
\alias{echoIBM.fishInside2}
\title{Discards fish outside of the sampling region of the sonar/echosounder.}
\usage{
echoIBM.extraxtRandSel(
  dynschool,
  data,
  rand.sel,
  dynschoolnames,
  staticschoolnames
)

echoIBM.fishInside(
  dynschool,
  data,
  dumpfile,
  discardOutside = c(r = Inf, az = Inf, el = Inf),
  rand.sel = 1
)

echoIBM.fishInside2(
  data,
  dumpfile,
  discardOutside = c(r = Inf, az = Inf, el = Inf),
  rand.sel = 1
)
}
\arguments{
\item{dynschool}{is a list of the dynamic fish information to be subsetted.}

\item{data}{is a list holding the variables 'esnm', and vessel spesifications.}
}
\description{
Discards fish outside of the sampling region of the sonar/echosounder.

Discards fish outside of the sampling region of the sonar/echosounder.

Discards fish outside of the sampling region of the sonar/echosounder FOR EACH SCHOOL.
}
\examples{
\dontrun{}

\dontrun{}

\dontrun{}

}
