% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.getFileTypes.R
\name{echoIBM.getFileTypes}
\alias{echoIBM.getFileTypes}
\title{Reads the header of all files in an event and classifies into the following file types: beams (beam configuration), ctd, vessel, pings, school, tsd, other.}
\usage{
echoIBM.getFileTypes(
  files,
  ext = list(beams = "beams", ctd = "ctd", pings = "pings", school = "school", vessel =
    "vessel", noise = NULL),
  labl = list(beams = c("cali"), noise = c("bgns", "nrns", "nr0a", "nr0p")),
  recursive = TRUE
)
}
\arguments{
\item{files}{list of files for which file type should be detemined.}

\item{ext}{Vector of file extensions to test the files against.}

\item{labl}{Key varialbe names corresponding to the file types given in \code{ext}.}

\item{recursive}{Used in list.files() when files is a single directory.}
}
\description{
Reads the header of all files in an event and classifies into the following file types: beams (beam configuration), ctd, vessel, pings, school, tsd, other.
}
\examples{
\dontrun{}

}
