% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoIBM.oneping.j2.R
\name{echoIBM.oneping.j2}
\alias{echoIBM.oneping.j2}
\title{Simulates one radial sphere (sample interval) 'j' of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. Two for loops used (for each frequency and each beam).}
\usage{
echoIBM.oneping.j2(j, data)
}
\arguments{
\item{j}{is the sampling interval index.}

\item{data}{is the list of data for the simulation, containing all of the following elements:}
}
\value{

}
\description{
Simulates one radial sphere (sample interval) 'j' of one echo sounder observation based on positions, orientations, sizes and other specifics of each fish in a known (simulated) school. Two for loops used (for each frequency and each beam).
}
\examples{
\dontrun{}

}
